ai-control-plane/requirements.txt
docker==7.1.0
psutil==6.1.1
redis==5.2.1
requests==2.32.3
prometheus-api-client==0.5.6
pyyaml==6.0.2
pydantic==2.10.4
tenacity==9.0.0
